New Classes added
New bean class Acknowledgment is added for sending response to client
when some data is inserted into data. This response is generated by
database and then it is transferred to fron with help of bean.

New method added in general service for supporting response transfer to
client.

New type of Exception is added UserNotFoundException

Some URL mappings are changed for using formparam annotations

new Interface StoAcknow is added for supporting transfer of response to
client